Move scanners
