+++ apps/passport-client/components/screens/MultiChoiceScanScreen.tsx
@@ -8,8 +8,8 @@ import { useLaserScannerKeystrokeInput } from "../../src/appHooks";
 import { loadUsingLaserScanner } from "../../src/localstorage";
 import { maybeRedirect } from "../../src/util";
 import { H5, Spacer, TextCenter } from "../core";
-import { ScanditScanner } from "../core/ScanditScanner";
-import { StrichScanner } from "../core/StrichScanner";
+import { ScanditScanner } from "../core/scanners/ScanditScanner";
+import { StrichScanner } from "../core/scanners/StrichScanner";
 import { AppContainer } from "../shared/AppContainer";
 import { IndicateIfOffline } from "../shared/IndicateIfOffline";
 import Select from "../shared/Select";
